requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
psycopg2==2.9.5
requests==2.28.2
python-dotenv==0.15.0
```